+++ lib/get-property-info/svg/svg-attribute-hook.js
