+++ lib/get-property-info/svg/svg-namespaces.js
@@ -395,22 +395,15 @@ var SVG_PROPERTIES = {
     'zoomAndPan': DEFAULT_NAMESPACE
 };
 
-function thisIsSVGTag(tag) {
-    return SVG_ELEMENTS.hasOwnProperty(tag);
-}
+module.exports.SVG_NAMESPACE = SVG_NAMESPACE;
+module.exports.SVG_ELEMENTS = SVG_ELEMENTS;
+module.exports.SVG_PROPERTIES = SVG_PROPERTIES;
 
-function getSVGNamespace() {
-    return SVG_NAMESPACE;
-}
 
-function SVGAttributeNamespace(value) {
-    if (SVG_PROPERTIES.hasOwnProperty(value)) {
-        return SVG_PROPERTIES[value];
-    }
-}
 
-module.exports.thisIsSVGTag = thisIsSVGTag;
 
-module.exports.getSVGNamespace = getSVGNamespace;
 
-module.exports.SVGAttributeNamespace = SVGAttributeNamespace;
+
+
+
+
