+++ lib/htmlparser-to-vdom.js
@@ -1,9 +1,7 @@
 var decode = require('ent').decode;
 var convertTagAttributes = require('./convert-tag-attributes');
-var thisIsSVGTag = require('./svg-namespaces').thisIsSVGTag,
-    getSVGNamespace = require('./svg-namespaces').getSVGNamespace,
-    SVGAttributeNamespace = require('./svg-namespaces').SVGAttributeNamespace,
-    SVGAttributeHook = require('./svg-attribute-hook');
+var svg = require('./get-property-info/svg');
+var SVGAttributeHook = require('./get-property-info/svg/svg-attribute-hook');
 
 module.exports = function createConverter (VNode, VText) {
     var converter = {
@@ -29,7 +27,7 @@ module.exports = function createConverter (VNode, VText) {
                 return converter.convert(node, getVNodeKey);
             });
 
-            if(thisIsSVGTag(tag.name)) {
+            if(svg.thisIsSVGTag(tag.name)) {
                 return convertSvg(tag, attributes, children, key);
             }
 
@@ -45,7 +43,7 @@ module.exports = function createConverter (VNode, VText) {
                 continue;
             }
 
-            var namespace = SVGAttributeNamespace(_key);
+            var namespace = svg.SVGAttributeNamespace(_key);
 
             if (namespace === void 0) { // not a svg attribute
                 continue;
@@ -67,7 +65,7 @@ module.exports = function createConverter (VNode, VText) {
             }
         }
 
-        return new VNode(tag.name, attributes, children, key, getSVGNamespace());
+        return new VNode(tag.name, attributes, children, key, svg.getSVGNamespace());
     }
